Moved source files into appropriate directory structure
