+++ src/main/java/dungeonboard/Mode.java
@@ -1,4 +1,4 @@
-package main;
+package dungeonboard;
 
 public enum Mode {
 	LAYER,
