+++ src/main/java/dungeonboard/Settings.java
@@ -1,4 +1,4 @@
-package main;
+package dungeonboard;
 
 import java.awt.Color;
 import java.awt.Dimension;
