+++ src/main/java/dungeonboard/display/Cube.java
@@ -1,11 +1,11 @@
-package display;
+package dungeonboard.display;
 
 import java.awt.Dimension;
 import java.awt.Graphics2D;
 import java.awt.Point;
 import java.util.Random;
 
-import main.Settings;
+import dungeonboard.Settings;
 
 public class Cube {
