+++ src/main/java/dungeonboard/display/Scale.java
@@ -1,4 +1,4 @@
-package display;
+package dungeonboard.display;
 
 public enum Scale {
 	FILL,
