+++ src/main/java/dungeonboard/paint/DrawMode.java
@@ -1,4 +1,4 @@
-package paint;
+package dungeonboard.paint;
 
 public enum DrawMode {
 	ANY,
