+++ src/main/java/dungeonboard/paint/Pen.java
@@ -1,4 +1,4 @@
-package paint;
+package dungeonboard.paint;
 
 public enum Pen {
 	CIRCLE,
