+++ src/main/resources/circle.gif
