+++ src/main/resources/hand0.png
