+++ src/main/resources/hand2.png
