+++ src/main/resources/hand3.png
