+++ src/main/resources/horizontal.gif
