+++ src/main/resources/mouse.gif
