+++ src/main/resources/square.gif
