+++ src/main/resources/squigle.gif
