+++ src/main/resources/vertical.gif
